backend/requirements.txt
# Dependencias para FastAPI y SQL
fastapi>=0.95.0,<0.96.0
pydantic>=1.10.7,<2.0.0
uvicorn>=0.21.1,<0.22.0
python-multipart>=0.0.6,<0.1.0
email-validator>=2.0.0,<3.0.0
passlib>=1.7.4,<2.0.0
python-jose>=3.3.0,<4.0.0
bcrypt>=4.0.1,<5.0.0
sqlalchemy>=2.0.9,<3.0.0
alembic>=1.10.3,<2.0.0
asyncpg>=0.27.0,<0.28.0
psycopg2-binary>=2.9.6,<3.0.0

# Dependencias para Google Calendar
google-auth>=2.17.3,<3.0.0
google-auth-oauthlib>=1.0.0,<2.0.0
google-auth-httplib2>=0.1.0,<0.2.0
google-api-python-client>=2.86.0,<3.0.0
